refactor: dedupe types and move out multisig
