+++ packages/near-api-js/src/utils/multisig.ts
@@ -1,0 +1,46 @@
+import { BN } from 'bn.js';
+import { parseNearAmount } from './format';
+
+export const MULTISIG_ALLOWANCE = new BN(parseNearAmount('1') as string);
+
+export const MULTISIG_CHANGE_METHODS = ['add_request', 'add_request_and_confirm', 'delete_request', 'confirm'];
+
+export const MULTISIG_GAS = new BN('100000000000000');
+
+export const MULTISIG_DEPOSIT = new BN('0');
+
+export const convertPublicKeyForMultisig = (pk) => pk.toString().replace('ed25519:', '');
+
+export const convertActionsForMultisig = (actions, accountId, receiverId) => actions.map((a) => {
+    const type = a.enum;
+    const { gas, publicKey, methodName, args, deposit, accessKey, code } = a[type];
+    const action: any = {
+        type: type[0].toUpperCase() + type.substr(1),
+        gas: (gas && gas.toString()) || undefined,
+        public_key: (publicKey && convertPublicKeyForMultisig(publicKey)) || undefined,
+        method_name: methodName,
+        args: (args && Buffer.from(args).toString('base64')) || undefined,
+        code: (code && Buffer.from(code).toString('base64')) || undefined,
+        amount: (deposit && deposit.toString()) || undefined,
+        deposit: (deposit && deposit.toString()) || '0',
+        permission: undefined,
+    };
+    if (accessKey) {
+        if (receiverId === accountId && accessKey.permission.enum !== 'fullAccess') {
+            action.permission = {
+                receiver_id: accountId,
+                allowance: MULTISIG_ALLOWANCE.toString(),
+                method_names: MULTISIG_CHANGE_METHODS,
+            };
+        }
+        if (accessKey.permission.enum === 'functionCall') {
+            const { receiverId: receiver_id, methodNames: method_names, allowance } = accessKey.permission.functionCall;
+            action.permission = {
+                receiver_id,
+                allowance: (allowance && allowance.toString()) || undefined,
+                method_names
+            };
+        }
+    }
+    return action;
+});
